Modifiche varie ai file.

+++ WineAdvisor/src/main/java/com/wineadvisor/wineadvisor/model/WineId.java
@@ -1,11 +1,12 @@
 package com.wineadvisor.wineadvisor.model;
 
+
 import lombok.Data;
 
 @Data
 public class WineId {
-    private int id;
+    private Long id;
     private String name;
-    private int year;
+    private int year; 
     private String image;
 }
